refactor(proto): share toml config edit helpers
