+++ nodelite-proto/Cargo.toml
@@ -14,4 +14,5 @@ ipnet.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 toml.workspace = true
+toml_edit.workspace = true
 url.workspace = true
