+++ nodelite-server/tests/fixtures/config_edit/alerts_base.toml
@@ -1,0 +1,10 @@
+[server]
+listen = "127.0.0.1:8080"
+public_base_url = "https://monitor.example.com"
+
+[alerts]
+enabled = false
+
+[auth]
+username = "viewer"
+password = "old-pass"
