+++ nodelite-server/tests/fixtures/config_edit/alerts_with_comments.toml
@@ -1,0 +1,14 @@
+[server]
+listen = "127.0.0.1:8080"
+public_base_url = "https://monitor.example.com"
+
+[alerts]
+enabled = false # keep enabled comment
+
+[alerts.smtp]
+enabled = false
+host = "old.example.com" # keep host comment
+
+[auth]
+username = "viewer"
+password = "old-pass"
