+++ nodelite-server/tests/fixtures/config_edit/auth_2fa_base.toml
@@ -1,0 +1,10 @@
+[server]
+listen = "127.0.0.1:8080"
+public_base_url = "https://monitor.example.com"
+
+[auth]
+username = "viewer"
+password = "old-pass"
+
+[ui]
+refresh_interval_secs = 5
