+++ nodelite-server/tests/fixtures/config_edit/auth_with_trailing_comment.toml
@@ -1,0 +1,13 @@
+[server]
+listen = "127.0.0.1:8080"
+public_base_url = "https://monitor.example.com"
+
+[auth]
+username = "viewer"
+password = "old-pass" # keep this comment
+
+[ui]
+welcome = """
+hello
+world
+"""
