+++ nodelite-server/tests/fixtures/config_edit/missing_auth.toml
@@ -1,0 +1,2 @@
+[server]
+listen = "127.0.0.1:8080"
